+++ Plugins/DynamicRefs/Source/NexusDynamicRefs/Private/NDynamicRefListViewEntry.cpp
@@ -1,0 +1,19 @@
+﻿#include "NDynamicRefListViewEntry.h"
+
+void UNDynamicRefListViewEntry::NativeDestruct()
+{
+	Super::NativeDestruct();
+}
+
+void UNDynamicRefListViewEntry::NativeOnListItemObjectSet(UObject* ListItemObject)
+{
+	INListViewEntry::NativeOnListItemObjectSet(ListItemObject);
+}
+
+void UNDynamicRefListViewEntry::Refresh() const
+{
+}
+
+void UNDynamicRefListViewEntry::Reset() const
+{
+}